% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxTest_supportingFunctions.R
\name{boot_optimization_function}
\alias{boot_optimization_function}
\title{Finding the restricted placebo coefficients for the maximum equivalence test based on the bootstrap approaches}
\usage{
boot_optimization_function(x, y, no_placebos, equiv_threshold, start_val)
}
\arguments{
\item{x}{The double demeaned independent variables.}

\item{y}{The double demeaned dependent variable.}

\item{no_placebos}{The number of placebo coefficients.}

\item{equiv_threshold}{The equivalence threshold for the test.}

\item{start_val}{The starting values for the optimization.}
}
\value{
A numeric vector containing the restricted placebo coefficients
}
\description{
\code{boot_optimization_function} solves the optimization problem to find the restricted placebo coefficients, according to Dette & Schumann (2024).
}
\references{
Dette, H., & Schumann, M. (2024). "Testing for Equivalence of Pre-Trends in Difference-in-Differences Estimation." \emph{Journal of Business & Economic Statistics}, 1–13. DOI: \doi{10.1080/07350015.2024.2308121}
}
