% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxTest_supportingFunctions.R
\name{double.demean}
\alias{double.demean}
\title{Double Demeaning a data.frame object}
\usage{
double.demean(df)
}
\arguments{
\item{df}{The data.frame object to double-demean. It should contain a column with names "ID" and "period" representing the unit and time period, respectively, over which the data is double-demeaned.}
}
\value{
a data.frame object with the double demeaned data.
}
\description{
Double Demeaning a data.frame object
}
