% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structuring.R
\name{EquiTrends_dataconstr}
\alias{EquiTrends_dataconstr}
\title{Data Construction Function for EquiTrends}
\usage{
EquiTrends_dataconstr(
  Y,
  ID,
  G,
  period,
  X,
  data,
  pretreatment_period,
  base_period,
  cluster
)
}
\arguments{
\item{Y}{see \link[EquiTrends]{maxEquivTest}, \link[EquiTrends]{meanEquivTest} or \link[EquiTrends]{rmsEquivTest}}

\item{ID}{see \link[EquiTrends]{maxEquivTest}, \link[EquiTrends]{meanEquivTest} or \link[EquiTrends]{rmsEquivTest}}

\item{G}{see \link[EquiTrends]{maxEquivTest}, \link[EquiTrends]{meanEquivTest} or \link[EquiTrends]{rmsEquivTest}}

\item{period}{see \link[EquiTrends]{maxEquivTest}, \link[EquiTrends]{meanEquivTest} or \link[EquiTrends]{rmsEquivTest}}

\item{X}{see \link[EquiTrends]{maxEquivTest}, \link[EquiTrends]{meanEquivTest} or \link[EquiTrends]{rmsEquivTest}}

\item{data}{see \link[EquiTrends]{maxEquivTest}, \link[EquiTrends]{meanEquivTest} or \link[EquiTrends]{rmsEquivTest}}

\item{pretreatment_period}{see \link[EquiTrends]{maxEquivTest}, \link[EquiTrends]{meanEquivTest} or \link[EquiTrends]{rmsEquivTest}}

\item{base_period}{see \link[EquiTrends]{maxEquivTest}, \link[EquiTrends]{meanEquivTest} or \link[EquiTrends]{rmsEquivTest}}

\item{cluster}{see \link[EquiTrends]{maxEquivTest}, \link[EquiTrends]{meanEquivTest} or \link[EquiTrends]{rmsEquivTest}}
}
\value{
A list containing the structured data.frame object used in the equivalence testing procedures, the base period for the test, a logical value indicating whether the panel is balanced and the number of periods.
}
\description{
Data Construction Function for EquiTrends
}
