% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-structuring.R
\name{EquiTrends_inputcheck}
\alias{EquiTrends_inputcheck}
\title{Input Checks Function for EquiTrends}
\usage{
EquiTrends_inputcheck(
  Y,
  ID,
  G,
  period,
  X,
  data,
  equiv_threshold,
  pretreatment_period,
  base_period,
  cluster,
  alpha
)
}
\arguments{
\item{Y}{see \link[EquiTrends]{maxEquivTest}, \link[EquiTrends]{meanEquivTest} or \link[EquiTrends]{rmsEquivTest}}

\item{ID}{see \link[EquiTrends]{maxEquivTest}, \link[EquiTrends]{meanEquivTest} or \link[EquiTrends]{rmsEquivTest}}

\item{G}{see \link[EquiTrends]{maxEquivTest}, \link[EquiTrends]{meanEquivTest} or \link[EquiTrends]{rmsEquivTest}}

\item{period}{see \link[EquiTrends]{maxEquivTest}, \link[EquiTrends]{meanEquivTest} or \link[EquiTrends]{rmsEquivTest}}

\item{X}{see \link[EquiTrends]{maxEquivTest}, \link[EquiTrends]{meanEquivTest} or \link[EquiTrends]{rmsEquivTest}}

\item{data}{see \link[EquiTrends]{maxEquivTest}, \link[EquiTrends]{meanEquivTest} or \link[EquiTrends]{rmsEquivTest}}

\item{equiv_threshold}{see \link[EquiTrends]{maxEquivTest}, \link[EquiTrends]{meanEquivTest} or \link[EquiTrends]{rmsEquivTest}}

\item{pretreatment_period}{see \link[EquiTrends]{maxEquivTest}, \link[EquiTrends]{meanEquivTest} or \link[EquiTrends]{rmsEquivTest}}

\item{base_period}{see \link[EquiTrends]{maxEquivTest}, \link[EquiTrends]{meanEquivTest} or \link[EquiTrends]{rmsEquivTest}}

\item{cluster}{see \link[EquiTrends]{maxEquivTest}, \link[EquiTrends]{meanEquivTest} or \link[EquiTrends]{rmsEquivTest}}

\item{alpha}{see \link[EquiTrends]{maxEquivTest}, \link[EquiTrends]{meanEquivTest} or \link[EquiTrends]{rmsEquivTest}}
}
\value{
A list containing an error indicator and a message. If \code{error} is TRUE, \code{message} contains an error message. If \code{error} is FALSE, \code{message} is empty.
}
\description{
Input Checks Function for EquiTrends
}
