% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxTest_supportingFunctions.R
\name{maxTestIU}
\alias{maxTestIU}
\title{maxTestIU}
\usage{
maxTestIU(data, delta, vcov, cluster, alpha, n, no.periods, base.period)
}
\arguments{
\item{data}{the dataframe containing the data from the dataconstruction function}

\item{delta}{the equivalence threshold}

\item{vcov}{the type of variance-covariance matrix to be used}

\item{cluster}{the cluster variable}

\item{alpha}{the significance level}

\item{n}{the number of individuals}

\item{no.periods}{the number of periods}

\item{base.period}{the base period}
}
\value{
a list containing the results of the maxTestIU function
}
\description{
maxTestIU
}
