% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanTest_supportingFunctions.R
\name{meanTest_optim_func}
\alias{meanTest_optim_func}
\title{Finding the minimum equivalence threshold for the mean equivalence test}
\usage{
meanTest_optim_func(coef, sd, alpha)
}
\arguments{
\item{coef}{The estimated absolute value of the mean placebo coefficients}

\item{sd}{The estimated standard deviation of the mean of the placebo coefficients}

\item{alpha}{The significance level}
}
\value{
The minimum equivalence threshold for which the null hypothesis of non-negligible differences can be rejected for the equivalence test based on the mean placebo coefficient.
}
\description{
\code{meanTest_optim_func} solves the optimization problem to find the minimum equivalence threshold for which one can reject the null hypothesis of non-negligible pre-trend differences at a given significance level for the equivalence test based on the mean placebo coefficient.
}
