% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{print.maxEquivTestIU}
\alias{print.maxEquivTestIU}
\title{Print maxEquivTestIU objects}
\usage{
\method{print}{maxEquivTestIU}(x, ...)
}
\arguments{
\item{x}{An object of class 'maxEquivTestIU' containing the results of the maximum test based on the intersection-union approach.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The function prints a summary of the results of the maximum test based on the intersection-union approach.
}
\description{
Print method for objects of class 'maxEquivTestIU'.
}
