% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{print.maxTestBoot}
\alias{print.maxTestBoot}
\title{Title}
\usage{
\method{print}{maxTestBoot}(x)
}
\arguments{
\item{x}{An object of class 'maxTestBoot' containing the results of the maximum test based on the bootstrap procedure.}
}
\value{
The function prints a summary of the results of the maximum test based on the bootstrap procedures.
}
\description{
Title
}
