% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-functions.R
\name{print.maxTestIU}
\alias{print.maxTestIU}
\title{Print maxTestIU objects}
\usage{
\method{print}{maxTestIU}(x)
}
\arguments{
\item{x}{An object of class 'maxTestIU' containing the results of the maximum test based on the intersection-union approach.}
}
\value{
The function prints a summary of the results of the maximum test based on the intersection-union approach.
}
\description{
Print method for objects of class 'maxTestIU'.
}
