% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{print.meanEquivTest}
\alias{print.meanEquivTest}
\title{Print meanEquivTest objects}
\usage{
\method{print}{meanEquivTest}(x, ...)
}
\arguments{
\item{x}{An object of class 'meanEquivTest' containing the results of the maximum test based on the bootstrap procedure.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The function prints a summary of the results of the maximum test based on the bootstrap procedures.
}
\description{
Print meanEquivTest objects
}
