% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{print.rmsEquivTest}
\alias{print.rmsEquivTest}
\title{Print rmsEquivTest objects}
\usage{
\method{print}{rmsEquivTest}(x, ...)
}
\arguments{
\item{x}{An object of class 'rmsEquivTest' containing the results of the maximum test based on the bootstrap procedure.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The function prints a summary of the results of the maximum test based on the bootstrap procedures.
}
\description{
Print rmsEquivTest objects
}
