% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxTest_supportingFunctions.R
\name{sigma_hathat_c}
\alias{sigma_hathat_c}
\title{Calculating the constrained variance of the residuals for the Boostrap approaches in the EquiTrends Maximum Equivalence Testing procedure, according to Dette & Schumann (2024).}
\usage{
sigma_hathat_c(parameter, x, y, ID, time)
}
\arguments{
\item{parameter}{The constrained coefficients.}

\item{x}{The double demeaned independent variables.}

\item{y}{The double demeaned dependent variable.}

\item{ID}{The ID variable.}

\item{time}{The time variable.}
}
\value{
The estimated constrained variance of the residuals.
}
\description{
Calculating the constrained variance of the residuals for the Boostrap approaches in the EquiTrends Maximum Equivalence Testing procedure, according to Dette & Schumann (2024).
}
\references{
Dette, H., & Schumann, M. (2024). "Testing for Equivalence of Pre-Trends in Difference-in-Differences Estimation." \emph{Journal of Business & Economic Statistics}, 1–13. DOI: \doi{10.1080/07350015.2024.2308121}
}
