% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsTest_supportingFunctions.R
\name{W_critical_value}
\alias{W_critical_value}
\title{Calculating the critical value for the W distribution as construced in Dette & Schumann (2024).}
\usage{
W_critical_value(significance_level)
}
\arguments{
\item{significance_level}{The significance level for the test. Must be one of 0.01, 0.025, 0.05, 0.1, 0.2, 0.8, 0.9, 0.95, 0.975, 0.99.}
}
\value{
A numeric scalar with the critical value for the W distribution at the given significance level.
}
\description{
Calculating the critical value for the W distribution as construced in Dette & Schumann (2024).
}
\references{
Dette, H., & Schumann, M. (2024). "Testing for Equivalence of Pre-Trends in Difference-in-Differences Estimation." \emph{Journal of Business & Economic Statistics}, 1–13. DOI: \doi{10.1080/07350015.2024.2308121}
}
