% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EquiTrends_Main_Functions.R
\name{meanEquivTest}
\alias{meanEquivTest}
\title{Equivalence Test for Pre-trends based on the Mean Placebo Coefficient}
\usage{
meanEquivTest(
  Y,
  ID,
  G,
  period,
  X = NULL,
  data = NULL,
  delta = NULL,
  pretreatment.period = NULL,
  base.period = NULL,
  vcov = NULL,
  cluster = NULL,
  alpha = 0.05,
  verbose = TRUE
)
}
\arguments{
\item{Y}{If 'data' is supplied, a scalar identifying the column number or column-name character string that corresponds to the numeric dependent (outcome) variable in ’data’. If 'data' is not supplied, a numeric vector with the variable of interest.}

\item{ID}{If 'data' is supplied, a scalar identifying the column number or column-name character string that corresponds to the unit numbers in ’data’. If 'data' is not supplied, a numeric vector (of the same dimension as Y) containing the unit numbers of the observations.}

\item{G}{If 'data' is supplied, a scalar identifying the column number or column-name character string associated to the binary or logic variable indicating if the individual receives treatment (e.g. 1 or TRUE) or not (0 or FALSE). If 'data' is not supplied, a vector (of the same dimension as Y) binary or logic indicating if the individual (e.g. the ID vector).}

\item{period}{If 'data' is supplied, a scalar identifying the column number or column-name character string associated with period (time) data. The time variable has to be numeric. If 'data' is not supplied, a numeric vector (of the same dimension as Y) indicating time.}

\item{X}{If 'data' is supplied, a vector of column numbers or column-name character strings that identifies the control variables’ columns. If data is not supplied, a vector, matrix or data.frame containing the control variables.}

\item{data}{An optional data.frame object containing the variables in Y, ID, G, T and, if supplied, X and cluster as its columns.}

\item{delta}{The scalar equivalence threshold (must be positive). The default is NULL, implying that the function must look for the minimum value for which the null of ”non-negligible differences” can still be rejected.}

\item{pretreatment.period}{A numeric vector identifying the pre-treatment periods that should be used for testing. The default is to use all periods that are included in T.}

\item{base.period}{The pre-treatment period to compare the post-treatment observation to. The default is to take the last specified pre-treatment period.}

\item{vcov}{The variance-covariance matrix that needs to be used. See details for more details.}

\item{cluster}{If vcov = "CL", a vector indicating which observations belong to the same cluster of the same length as Y. If 'data' is supplied, 'cluster' must be either the column index or column name of this vector in the data.frame/matrix. The default (cluster=NULL) assumes every unit in ID is its own cluster.}

\item{alpha}{Significance level of the test. The default is 0.05.}

\item{verbose}{A logical object indicating if test results need to be printed. The default is TRUE.}
}
\value{
hoi
}
\description{
This function performs an equivalence test for pre-trends based on the maximum absolute placebo coefficient. The test can be performed using the intersection-union approach (IU), a bootstrap procedure for spherical errors (Boot) and a wild bootstrap procedure (Wild).
}
