% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EquiTrends_Main_Functions.R
\name{meanEquivTest}
\alias{meanEquivTest}
\title{Equivalence Test for Pre-trends based on the Mean Placebo Coefficient}
\usage{
meanEquivTest(
  Y,
  ID,
  G,
  period,
  X = NULL,
  data = NULL,
  equiv_threshold = NULL,
  pretreatment_period = NULL,
  base_period = NULL,
  vcov = NULL,
  cluster = NULL,
  alpha = 0.05
)
}
\arguments{
\item{Y}{A numeric vector with the variable of interest. If \code{data} is supplied, \code{Y} should be a scalar indicating the column number or column-name character string that corresponds to the numeric dependent (outcome) variable in ’data’.}

\item{ID}{A numeric vector identifying the different cross-sectional units in the dataset. If \code{data} is supplied, \code{ID} should be a scalar indicating the column number or column-name character string that corresponds to the cross-sectional units identifier in \code{data}.}

\item{G}{A binary or logic vector (of the same dimension as \code{Y} and \code{ID}) indicating if the individual (e.g. as indicated by \code{ID}) receives treatment (e.g. 1 or TRUE) or not (0 or FALSE). f 'data' is supplied, \code{G} should be a scalar identifying the column number or column-name character string associated to \code{G} in \code{data}.}

\item{period}{A numeric vector (of the same dimension as Y) indicating time. If \code{data} is supplied, \code{period} should be a scalar indicating the column number or column-name character string that corresponds to the time identifier in \code{data}.}

\item{X}{A vector, matrix, or data.frame containing the control variables. If \code{data} is supplied, \code{X} must be a vector of column numbers or column-name character strings that identifies the control variables’ columns.}

\item{data}{An optional \code{data.frame} object containing the variables in Y, ID, G, T and, if supplied, X and cluster as its columns.}

\item{equiv_threshold}{The scalar equivalence threshold (must be positive). The default is NULL, implying that the function must look for the minimum value for which the null hypothesis of ”non-negligible differences” can still be rejected.}

\item{pretreatment_period}{A numeric vector identifying the pre-treatment periods that should be used for testing. \code{pretreatment_period} must be a subset of the periods included through \code{period}. The default is to use all periods that are included in \code{period}.}

\item{base_period}{The pre-treatment period to compare the post-treatment observation to. The default is to take the last period of the pre-treatment period.}

\item{vcov}{The variance-covariance matrix that needs to be used. See \emph{Details} for more details.}

\item{cluster}{If \code{vcov = "CL"}, a vector indicating which observations belong to the same cluster. \code{cluster} must be of the same length as the panel. If \code{data} is supplied, \code{cluster} must be either the column index or column name of this vector in the data.frame/matrix. The default (\code{cluster=NULL}) assumes every unit in ID is its own cluster.}

\item{alpha}{Significance level of the test. The default is 0.05. Only required if \code{equiv_threshold} is not specified.}
}
\value{
An object of class "meanEquivTest" containing:
\item{\code{placebo_coefficients}}{ a numeric vector of the estimated placebo coefficients,}
\item{\code{abs_mean_placebo_coefs}}{ the absolute value of the mean of the placebo coefficients,}
\item{\code{var_mean_placebo_coef}}{ the estimated variance of the mean placebo coefficient,}
\item{\code{significance_level}}{ the significance level of the test,}
\item{\code{base_period}}{ the base period used in the testing procedure,}
\item{\code{num_individuals}}{ the number of cross-sectional individuals in the panel used for testing,}
\item{\code{num_periods}}{ the number of periods in the panel used for testing (if the panel is unbalanced, \code{num_periods} represents the range in the number of time periods covered by different individuals),}
\item{\code{num_observations}}{ the total number of observations in the panel used for testing,}
\item{\code{is_panel_balanced}}{ a logical value indicating whether the panel is balanced,}
\item{\code{equiv_threshold_specified}}{ a logical value indicating whether an equivalence threshold was specified.}

If \code{equiv_threshold_specified = FALSE}, then additionally \code{minimum_equiv_threshold}: the minimum equivalence threshold for which the null hypothesis of non-negligible (based on the equivalence threshold) trend-differences can be rejected. 

If \code{equiv_threshold_specified = TRUE}, then additionally \itemize{
\item \code{mean_critical_value}: the critical value at the alpha level,
\item \code{p_value}: the p-value of the test,
\item \code{reject_null_hypothesis}: A logical value indicating whether to reject the null hypothesis,
\item \code{equiv_threshold}: the equivalence threshold specified.
}
}
\description{
This function performs an equivalence test for pre-trends based on the mean placebo coefficient from Dette & Schumann (2024).
}
\details{
The \code{vcov} parameter specifies the variance-covariance matrix to be used in the function. 
This parameter can take two types of inputs:
\enumerate{
\item A character string specifying the type of variance-covariance matrix estimation. The options are:\itemize{
   \item \code{NULL}: The default variance-covariance matrix estimated by the \link[plm]{plm} function is used.
   \item \code{"HC"}: A heteroscedasticity-robust (HC) covariance matrix is estimated using the \code{vcovHC} function from the \code{plm} package, \link[plm]{vcovHC}, with type \code{"HC1"} and method \code{"white1"} (see White, 1980).
   \item \code{"HAC"}: A heteroscedasticity and autocorrelation robust (HAC) covariance matrix is estimated using the \code{vcovHC} function from the \code{plm} package, \link[plm]{vcovHC}, with type \code{"HC3"} and method \code{"arellano"} (see Arellano, 1987).
   \item \code{"CL"}: A cluster-robust covariance matrix is estimated using the \code{vcovCR} function from the \code{clubSandwich} package with type \code{"CR0"} (see Lian & Zegers (1986)). The cluster variable is either \code{"ID"} or a custom cluster variable provided in the \code{data} dataframe.
}
\item A function that takes an \link[plm]{plm} object as input and returns a variance-covariance matrix. 
   This allows for custom variance-covariance matrix estimation methods. For example, you could 
   use the \code{vcovHC} function from the \code{sandwich} package with a specific method and type:
   \preformatted{function(x) {vcovHC(x, method = "white1", type = "HC2")}}
}
If no \code{vcov} parameter is provided, the function defaults to using the variance-covariance matrix 
estimated by the \link[plm:plm]{plm::plm()} function.

One should note that rows containing \code{NA} values are removed from the panel before the testing procedure is performed.
}
\examples{
# Generate a balanced panel dataset with 500 cross-sectional units (individuals), 
# 5 time periods (labeled 1-5), a binary variable indicating which individual 
# receives treatment and 2 control variables ("X_1" and "X_2") 
# The error-terms are generated without heteroscedasticity, autocorrelation, 
# or any significant clusters. Furthermore, there are no fixed effects 
# and no pre-trends present in the data (all values in beta are 0). 
# See sim_paneldata() for more details.

sim_data <- sim_paneldata(N = 500, tt = 5, p = 2, beta = rep(0, 5), 
                          gamma = rep(1, 2), het = 0, phi = 0, sd = 1, 
                          burnins = 50)
                          
# Perform the test with equivalent threshold specified as 1 based on 
# pre-treatment periods 1-4 and assuming homoscedastic error-terms:
  # To select variables, one can use the column names / column numbers in the panel data:
  meanEquivTest(Y = "Y", ID = "ID", G = "G", period = 2, X = c(5, 6),
                data = sim_data, equiv_threshold = 1, pretreatment_period = 1:4,
                base_period = 4)
  # Alternatively, one can use separate variables:
  data_Y <- sim_data$Y
  data_ID <- sim_data$ID
  data_G <- sim_data$G
  data_period <- sim_data$period
  data_X <- sim_data[, c(5, 6)]
  meanEquivTest(Y = data_Y, ID = data_ID, G = data_G, period = data_period, X = data_X,
                equiv_threshold = 1, pretreatment_period = 1:4,
                base_period = 4)
   
# Perform the test with a heteroscedastic and autocorrelation robust 
# variance-covariance matrix estimator, and without specifying the equivalence threshold:
meanEquivTest(Y = "Y", ID = "ID", G = "G", period = "period", X = c(5, 6),
              data = sim_data, equiv_threshold = NULL, pretreatment_period = 1:4,
              base_period = 4, vcov = "HAC")

# Perform the test with an equivalence threshold of 1 and a custom
# variance-covariance matrix estimator:
vcov_func <- function(x) {plm::vcovHC(x, method = "white1", type = "HC2")}
meanEquivTest(Y = "Y", ID = "ID", G = "G", period = "period", 
              data = sim_data, equiv_threshold = 1, pretreatment_period = 1:4,
              base_period = 4, vcov = vcov_func)
               
# Perform the test using clustered standard errors based on a vector indicating 
# the cluster. For instance, two clusters with the following rule: all
# individuals with an ID below 250 are in the same cluster:
cluster_ind <- ifelse(sim_data$ID < 250, 1, 2)
meanEquivTest(Y = data_Y, ID = data_ID, G = data_G, period = data_period, X = data_X,
               equiv_threshold = 1, pretreatment_period = 1:4,
               base_period = 4, vcov = "CL", cluster = cluster_ind)

# Note that the testing procedure can also handle unbalanced panels. 
# Finally, one should note that the test procedure also works for unbalanced panels.
# To illustrate this, we generate an unbalanced panel dataset by randomly selecting
# 70\% of the observations from the balanced panel dataset:
random_indeces <- sample(nrow(sim_data), 0.7*nrow(sim_data))
unbalanced_sim_data <- sim_data[random_indeces, ]
meanEquivTest(Y = "Y", ID = "ID", G = "G", period = "period", X = c(5, 6),
              data = unbalanced_sim_data, equiv_threshold = 1, pretreatment_period = 1:4,
              base_period = 4, vcov = "HAC")


}
\references{
Arellano M (1987). “Computing Robust Standard Errors for Within-groups Estimators.” \emph{Oxford bulletin of Economics and Statistics}, 49(4), 431–434.

Dette, H., & Schumann, M. (2024). "Testing for Equivalence of Pre-Trends in Difference-in-Differences Estimation." \emph{Journal of Business & Economic Statistics}, 1–13. DOI: \doi{10.1080/07350015.2024.2308121}

Liang, K.-Y., & Zeger, S. L. (1986). "Longitudinal data analysis using generalized linear models." \emph{Biometrika}, 73(1), 13-22. DOI: \doi{10.1093/biomet/73.1.13}

White H (1980). “A heteroskedasticity-consistent covariance matrix estimator and a direct test for heteroskedasticity.” \emph{Econometrica}, 48(4), 817–838.
}
\seealso{
\code{\link[=print.meanEquivTest]{print.meanEquivTest}}
}
\author{
Ties Bos
}
